Add files via upload
this function file is used to calculate antibody intensity from antibody only file. 
Note: For PE, 1-peak point is not working, due to ? (I am still thinking on this part)
So in this code, we are using 3-points to get good distribution and fitting for deconvolution analysis to get the copy number.